# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/subscriptions/:username": {
            "post": {
                "security": [
                    {
                        "UserKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "\"Подписка на другого пользователя\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username пользователя",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Все ок"
                    },
                    "500": {
                        "description": "Ошибка подписки"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "UserKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "\"Отписка от пользователя\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username пользователя",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Все ок"
                    },
                    "500": {
                        "description": "Ошибка отписки"
                    }
                }
            }
        },
        "/user/:username": {
            "get": {
                "security": [
                    {
                        "UserKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "\"Получение пользователя\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username пользователя",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Все ок"
                    },
                    "500": {
                        "description": "Ошибка получения"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "security": [
                    {
                        "UserKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "\"Получение текущего пользователя\"",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user_id пользователя",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Все ок"
                    },
                    "500": {
                        "description": "Ошибка разлогина"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "UserKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "\"Обновление инфо о пользователе\"",
                "parameters": [
                    {
                        "description": "Инфа пользователя",
                        "name": "info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Все ок"
                    },
                    "500": {
                        "description": "Ошибка обновления"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "UserKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "\"Создание пользователя\"",
                "parameters": [
                    {
                        "description": "Данные пользователя",
                        "name": "signapData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.signupData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.subsResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка введеных данных"
                    },
                    "500": {
                        "description": "Ошибка создания"
                    }
                }
            }
        },
        "/users/admin/:username": {
            "post": {
                "security": [
                    {
                        "UserKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "\"Логин пользователя\"",
                "parameters": [
                    {
                        "description": "Данные пользователя для логина",
                        "name": "loginData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.loginData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Все ок"
                    },
                    "500": {
                        "description": "Ошибка логина"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "UserKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "\"Разлогин пользователя\"",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "session_id пользователя",
                        "name": "session_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Все ок"
                    },
                    "500": {
                        "description": "Ошибка разлогина"
                    }
                }
            }
        }
    },
    "definitions": {
        "http.loginData": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "http.signupData": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "http.subsResponse": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "max_items": {
                    "type": "integer"
                },
                "pages_number": {
                    "type": "integer"
                },
                "subs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserNoPassword"
                    }
                }
            }
        },
        "models.Actor": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "biography": {
                    "type": "string"
                },
                "birthdate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_liked": {
                    "type": "boolean"
                },
                "movies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MovieReference"
                    }
                },
                "movies_count": {
                    "type": "integer"
                },
                "movies_rating": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                },
                "profession": {
                    "type": "string"
                }
            }
        },
...
        "models.MovieReference": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "favorite_actors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Actor"
                    }
                },
                "movies_watched": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "reviews_number": {
                    "type": "integer"
                },
                "subscribers": {
                    "type": "integer"
                },
                "subscriptions": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.UserNoPassword": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "favorite_actors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Actor"
                    }
                },
                "movies_watched": {
                    "type": "integer"
                },
                "reviews_number": {
                    "type": "integer"
                },
                "subscribers": {
                    "type": "integer"
                },
                "subscriptions": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}